+++ apps/passport-server/src/services.ts
@@ -13,8 +13,8 @@ import { startPersistentCacheService } from "./services/persistentCacheService";
 import { startPoapService } from "./services/poapService";
 import { startProvingService } from "./services/provingService";
 import { startRateLimitService } from "./services/rateLimitService";
-import { startRollbarService } from "./services/rollbarService";
 import { startSemaphoreService } from "./services/semaphoreService";
+import { startRollbarService } from "./services/startRollbarService";
 import { startTelegramService } from "./services/telegramService";
 import { startTelemetry } from "./services/telemetryService";
 import { startUserService } from "./services/userService";
